Fix data types

+++ nativelib/src/main/resources/scala-native/EmbeddedResourceReader.c
@@ -1,25 +1,26 @@
+#include <stdint.h>
 extern int** __resources_all_path;
 extern int** __resources_all_content;
-extern long* __resources_all_content_lengths;
-extern long* __resources_all_path_lengths;
-extern long __resources_size;
+extern int32_t* __resources_all_content_lengths;
+extern int32_t* __resources_all_path_lengths;
+extern int32_t __resources_size;
 
-int scalanative_resource_get_content_byte(int embeddedFileId, int byteIndex) {
+int scalanative_resource_get_content_byte(int32_t embeddedFileId, int32_t byteIndex) {
     return __resources_all_content[embeddedFileId][byteIndex];
 }
 
-int scalanative_resource_get_path_byte(int embeddedFileId, int byteIndex) {
+int scalanative_resource_get_path_byte(int32_t embeddedFileId, int32_t byteIndex) {
     return __resources_all_path[embeddedFileId][byteIndex];
 }
 
-long scalanative_resource_get_embedded_size() {
+int32_t scalanative_resource_get_embedded_size() {
     return __resources_size;
 }
 
-long scalanative_resource_get_path_length(int embeddedFileId) {
+int32_t scalanative_resource_get_path_length(int32_t embeddedFileId) {
     return __resources_all_path_lengths[embeddedFileId];
 }
 
-long scalanative_resource_get_content_length(int embeddedFileId) {
+int32_t scalanative_resource_get_content_length(int32_t embeddedFileId) {
     return __resources_all_content_lengths[embeddedFileId];
 }
